reste normes dans .h / game / save_load .
