+++ srcs/main.c
@@ -6,12 +6,26 @@
 /*   By: hsabouri <hsabouri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/14 18:07:18 by hsabouri          #+#    #+#             */
-/*   Updated: 2019/06/18 16:56:51 by hsabouri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <doom.h>
 #include <editor.h>
+
+static void		launch_game2(t_env *env)
+{
+	env->game.animations = safe_anew(NULL, 10, sizeof(t_animation), "init");
+	env->game.col_events = safe_anew(NULL, 10, sizeof(t_col_event), "init");
+	env->game.weapons = init_weapons(&env->game);
+	launch_check(env, env->game);
+	env->game.id_buf = (u_int32_t *)safe_malloc((WIDTH * HEIGHT
+	* sizeof(int)), "doom_nukem");
+	if (env->component)
+		destroy_component(env->component);
+	find_center_sectors(env->game);
+	env->component = init_component(env, &env->sdl);
+	env->init_game = 0;
+}
 
 static void		launch_game(t_env *env, int ac, char **av)
 {
@@ -36,17 +50,7 @@ static void		launch_game(t_env *env, int ac, char **av)
 		env->game = generate_map();
 		env->game = init_audio(env->game);
 	}
-	env->game.animations = safe_anew(NULL, 10, sizeof(t_animation), "init");
-	env->game.col_events = safe_anew(NULL, 10, sizeof(t_col_event), "init");
-	env->game.weapons = init_weapons(&env->game);
-	launch_check(env, env->game);
-	env->game.id_buf = (u_int32_t *)safe_malloc((WIDTH * HEIGHT
-		* sizeof(int)), "doom_nukem");
-	if (env->component)
-		destroy_component(env->component);
-	find_center_sectors(env->game);
-	env->component = init_component(env, &env->sdl);
-	env->init_game = 0;
+	launch_game2(env);
 }
 
 static void		all_loops(t_env *env, size_t frame)
