+++ srcs/save_load/parse_textures_sounds.c
@@ -3,10 +3,8 @@
 /*                                                        :::      ::::::::   */
 /*   parse_textures_sounds.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hugo <hugo@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/19 17:10:02 by iporsenn          #+#    #+#             */
-/*   Updated: 2019/06/06 12:52:05 by hugo             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
